pyproject.toml
[project]
name = "ai-agents-journal"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.118.3",
    "jinja2>=3.1.6",
    "openai-agents[litellm]>=0.3.3",
    "pip>=25.2",
    "python-dotenv>=1.1.1",
    "python-telegram-bot[job-queue]>=22.5",
    "ruff>=0.14.0",
    "tavily-python>=0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["config"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["agents"]
